requirements.txt
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
h11==0.14.0
idna==3.10
interchange==2021.0.4
monotonic==1.6
outcome==1.3.0.post0
packaging==24.2
pansi==2024.11.0
pillow==11.1.0
py2neo==2021.2.4
Pygments==2.19.1
pymodbus==3.8.6
PySocks==1.7.1
python-nmap==0.7.1
pytz==2025.1
requests==2.32.3
selenium==4.30.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
